Improve utils and components
